Added koala to the gadget making tool chain.
Update 3mm, 5mm, and 10mm LEDs to use koala for their resistors.  It's a start.

Fix to make jettest build again.

modified gadgetmaker2 to use HighEagle for file output (and to normalize the layers on the .brd and .sch files it generates)

Docs for EagleTools.py

Added function to EagleTools.py to propagate parts from .sch to .brd files.

Added function to EagleTools.py rebuild signals in brd based on sch.

fixed some bugs in HighEagle.jinja.py and GenerateHighEagle.py

Sort tags in output to make it feasible to diff files.

Made koala build a brd output in addition to an sch.





git-svn-id: svn+ssh://bbfs-01.calit2.net/grw/Gordon/svn/trunk/Gadgets/Tools/CircuitsByCode@20397 2cf53259-273a-af40-bda6-69d31023b8ec